If a comic is not in the database, insert before showing the bookmark list.
In the past versions of ComicViewerActivity, if the bookmark list was
shown and the comic was not in the database, global boormarks were shown
(not usable, since the referred comics were not loaded) Now, insert the
comic into the database before showing the bookmark list. In addition,
do not show this list if there are not bookmarks.

Some missing Spanish translations added.